MapScreenとdashboardの一部機能を統合
